Fix for null abstract properties
If a property's type is abstract and the value is set to null, BroadWorks still expects an xsi:type attribute to be set on the serialized XML for the property. This wasn't happening before. Now, an attribute will be set to the first concrete type found for the abstract type since it doesn't matter which type is used as long as it inherits the property's abstract type.